Add files via upload
Alumni Records using Java and Oracle Database is designed for the colleges to store the details of the Alumni. Alumni are usually referred to a group of passed out students from the colleges. The Alumni Records is an application based project. It is designed for the colleges to store the details of the alumni of the colleges. The existing systems of storing the alumni records in copies and also built with innumerable excel sheets. This is not efficient and not good to do. The proposed system will provide registration procedure to get sign up and login procedure for the admin through which can add or search or list all or update or download the details of the alumni of the colleges. The students completed his studies and begin his professional guidelines for upcoming students. So this system will help the colleges as well as the students through the colleges to be in contact with the alumni. There is no module to delete the details of the alumni in this application based project because alumni would never be deleted. An Alumni Records is an application based project which will store the detailed information about alumni such as, USN, first name, last name, date of birth, gender, father’s name, department, email address, mobile number, etc.